pyproject.toml
[project]
name = "aq-agent"
version = "0.1.0"
description = "Analytics Query Agent that intelligently handles user requests for analytics reports"
authors = [
    {name = "srinivas-tanla",email = "srinivas.mannem@tanla.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "unstructured (>=0.16.20,<0.17.0)",
    "pdfminer-six (>=20240706,<20240707)",
    "weaviate-client (>=4.10.4,<5.0.0)",
    "transformers (>=4.48.3,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "groq (>=0.18.0,<0.19.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false